S06P21S004-419 [be] fix: 이미지 크롭 수정

Closes S06P21S004-419

See merge request s06-s-project/S06P21S004!166

+++ server/thirdeye/src/main/java/com/aeye/thirdeye/dto/ImageDto.java
@@ -20,10 +20,10 @@ public class ImageDto {
     private String Provider;
     private String faceYN;
 
-    private double L_X;
-    private double L_Y;
-    private double R_X;
-    private double R_Y;
+    private int L_X;
+    private int L_Y;
+    private int R_X;
+    private int R_Y;
 
     public ImageDto(Image image){
         this.id = image.getId();
